Improve code quality
Bin, TaskList and Storage classes: refactor for better readability
and maintainability

The method name was not following the naming conventions for methods.
Method name was changed from load() to loadingTasks().
The renaming to loadingTasks() method in Storage class allows better
readility.

The breaking of lines in TaskList class is not aligned with the
standards. The break for line 63 and 77 was after the '+'. It is changed
to before the '+'. This will ensure better readability.